# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is an API documentation for SriRejeki IoT Greenhouse Monitoring",
        "title": "SriRejeki Backend API Documentation",
        "contact": {
            "name": "API Support",
            "email": "mail@hilmy.dev"
        },
        "version": "1.0"
    },
    "host": "localhost:80",
    "basePath": "/",
    "paths": {
...
        "/data/snapshot": {
            "get": {
                "description": "Get microcontroller snapshot data using authorization bearer header",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Data"
                ],
                "summary": "Get microcontroller snapshot data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Type of microcontroller sensor",
                        "name": "type",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "deviceId/clientId of the microcontroller",
                        "name": "deviceId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Duration of snapshot data: hour | day | month. Default: month",
                        "name": "duration",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Limit of snapshot data. It must be string form of a number. Default: 30",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/data.dataResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/data.dataResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/data.dataResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/data.dataResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Post microcontroller snapshot data | INTERNAL USE ON BACKEND ONLY",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Data"
                ],
                "summary": "Post microcontroller snapshot data | INTERNAL USE ON BACKEND ONLY",
                "parameters": [
                    {
                        "description": "Required data to post microcontroller snapshot data",
                        "name": "dataSnapshotPostRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/data.dataSnapshotPostRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/data.dataResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/data.dataResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/data.dataResponse"
                        }
                    }
                }
            }
        },
        "/data/subscribe": {
            "get": {
                "description": "Get mqtt data in specified topic using authorization bearer header",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Data"
                ],
                "summary": "Get mqtt data in specified topic",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Mqtt topic",
                        "name": "topic",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/data.dataResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/data.dataResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/data.dataResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/data.dataResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "data.dataResponse": {
            "type": "object",
            "properties": {
                "payload": {}
            }
        },
        "data.dataSnapshotPostRequest": {
            "type": "object",
            "required": [
                "data",
                "deviceId",
                "durations",
                "type"
            ],
            "properties": {
                "data": {
                    "type": "string"
                },
                "deviceId": {
                    "type": "string"
                },
                "durations": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
    }
}